requirements.txt
python-logstash-async==2.3.0
opentelemetry-sdk==1.7.1
opentelemetry-exporter-jaeger-thrift==1.7.1
opentelemetry-propagator-jaeger==1.7.1

opentelemetry-instrumentation-logging==0.26b1
opentelemetry-instrumentation-falcon==0.26b1
opentelemetry-instrumentation-celery==0.26b1
